test runner improvments

+++ bin/composer.sh
@@ -1,1 +1,5 @@
-docker-compose run --rm composer $@
+SCRIPT_DIR=$( dirname "${BASH_SOURCE[0]}" )
+
+if [ -x "${SCRIPT_DIR}/depcheck.sh" ]; then
+    docker-compose run --rm composer $@
+fi
