+++ bin/phpunit.sh
@@ -1,1 +1,4 @@
-docker-compose run --rm wordpress-phpunit phpunit $@
+SCRIPT_DIR=$( dirname "${BASH_SOURCE[0]}" )
+if [ -x "${SCRIPT_DIR}/depcheck.sh" ]; then
+    docker-compose run --rm wordpress-phpunit phpunit $@
+fi
